% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsproto.R
\docType{data}
\name{background}
\alias{background}
\title{Background}
\format{An object of class \code{background} (inherits from \code{facade}, \code{dsproto}) of length 2.}
\usage{
background
}
\description{
Background
}
\keyword{internal}
