% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmodel.R
\name{boolpos}
\alias{boolpos}
\title{Find Positions of False Elements}
\usage{
boolpos(fun, v)
}
\arguments{
\item{fun}{A \code{function(x)} which outputs a single boolean}

\item{v}{A vector or list of elements compatible with \code{fun}}
}
\description{
Returns all positions of elements in \code{v} where \code{fun(v)}
returns \code{FALSE}.
}
\keyword{internal}
