% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelConstruction.R
\name{+.dsproto}
\alias{+.dsproto}
\alias{\%+\%}
\alias{addToModel}
\title{Adds a new component to a dsmodel.}
\usage{
e1 + e2

e1 \%+\% e2

addToModel(model, obj)
}
\arguments{
\item{e1}{A model or feature, usually a model.}

\item{e2}{A model or feature, usually a feature.}
}
\description{
This operator allows you to add objects to a model, or to collect objects before they are added
to a model.
}
\details{
The operation is fully commutative, so order does not matter.
Models that do not yet have ranges will generally save features until the range is provided.

Supported features include dsdots, dsarrows, dscurves, dspoints, and colorRegions
The \code{+} operator should be the only one required.
}
\keyword{internal}
