% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmodel.R
\name{dsassert}
\alias{dsassert}
\title{Asserts a boolean condition is true}
\usage{
dsassert(t, str, critical = FALSE)
}
\arguments{
\item{t}{Boolean condition which must be true.}

\item{str}{Error message to return if the condition fails.}

\item{critical}{If the condition can only arise from a bug in dsmodels.}
}
\description{
This function takes a boolean and a string, replicating the common assert pattern.
}
\keyword{internal}
