% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsproto.R
\docType{data}
\name{feature}
\alias{feature}
\title{Feature}
\format{An object of class \code{feature} (inherits from \code{facade}, \code{dsproto}) of length 2.}
\usage{
feature
}
\description{
Feature
}
\keyword{internal}
