% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{features}
\alias{features}
\title{Features}
\usage{
features(f1, f2)
}
\arguments{
\item{f1}{dsproto of type feature}

\item{f2}{dsproto of type feature}
}
\description{
Defines a category for the different kinds of non-model dsprotos
}
\keyword{internal}
