% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspoint.R
\name{findFixedPoint}
\alias{findFixedPoint}
\title{Finds which fixed-point a point will end up at under a certain function.}
\usage{
findFixedPoint(x, y, points, eps, tolerance, model, stride)
}
\arguments{
\item{x}{A numeric initial x coordinate.}

\item{y}{A numeric initial y coordinate.}

\item{points}{A list comprised of x coordinates, y coordinates, such as output by \code{pointsToList}. Should represent the attractors of \code{fun}.}

\item{eps}{An epsilon, expected to already be squared, used to determine when a point is closed enough to a fixed point.}

\item{tolerance}{A, usually smaller, value used to determine when a point has stopped moving without finding a nearby fixpoint.}

\item{model}{A dsmodel encapsulating the function to be applied.}

\item{stride}{The number of times to apply the function at each step.}
}
\description{
Finds which fixed-point a point will end up at under a certain function.
}
\keyword{internal}
