% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspoint.R
\name{findNearestPoint}
\alias{findNearestPoint}
\title{Determines which point in a list is closest to the input. 0 is used if all are further from  epsilon.}
\usage{
findNearestPoint(x, y, points, eps, deep = TRUE)
}
\arguments{
\item{x}{A numeric  x coordinate.}

\item{y}{A numeric  y coordinate.}

\item{points}{A list comprised of x coordinates, y coordinates, such as output by \code{pointsToList}. Should represent the attractors of \code{fun}.}

\item{eps}{An epsilon, expected to already be squared, used to determine when a point is close enough. Only important if \code{deep} is true.}

\item{deep}{If true, only return a point if it is within epsilon distance. Otherwise, return the closest point.}
}
\description{
Determines which point in a list is closest to the input. 0 is used if all are further from  epsilon.
}
\keyword{internal}
