% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmodel.R
\name{finite.points}
\alias{finite.points}
\title{Check if points are finite}
\usage{
finite.points(points)
}
\arguments{
\item{points}{a list of points to test}
}
\description{
This function takes a list of points and returns true if all values in points are finite.
}
\keyword{internal}
