% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsproto.R
\name{is.background}
\alias{is.background}
\title{Checks if object is a background}
\usage{
is.background(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Checks if object is a background
}
\keyword{internal}
