% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dscurve.R
\name{is.curve}
\alias{is.curve}
\title{Reports whether x is a dscurves object.}
\usage{
is.curve(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether x is a dscurves object.
}
\keyword{internal}
