% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrange.R
\name{is.discretizedrange}
\alias{is.discretizedrange}
\title{Reports whether x is a discretized range.}
\usage{
is.discretizedrange(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether x is a discretized range.
}
\keyword{internal}
