% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdots.R
\name{is.dots}
\alias{is.dots}
\title{Reports whether x is a dsdots object.}
\usage{
is.dots(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether x is a dsdots object.
}
\keyword{internal}
