% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsarrows.R
\name{is.dsarrows}
\alias{is.dsarrows}
\title{Reports whether x is a dsarrows object.}
\usage{
is.dsarrows(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether x is a dsarrows object.
}
\keyword{internal}
