% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssimulation.R
\name{is.dsimage}
\alias{is.dsimage}
\title{Reports whether x is a dsimage}
\usage{
is.dsimage(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether x is a dsimage
}
\keyword{internal}
