% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsproto.R
\name{is.dsproto}
\alias{is.dsproto}
\title{Tests if an object is a ds object.}
\usage{
is.dsproto(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Tests if an object is a ds object.
}
\keyword{internal}
