% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsregion.R
\name{is.dsregion}
\alias{is.dsregion}
\title{Reports whether x is a dsregion object.}
\usage{
is.dsregion(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether x is a dsregion object.
}
\keyword{internal}
