% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{is.features}
\alias{is.features}
\title{Checks if object is a collection of dsproto objects.}
\usage{
is.features(obj)
}
\arguments{
\item{obj}{An object to test}
}
\description{
Checks if object is a collection of dsproto objects.
}
\keyword{internal}
