% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsproto.R
\name{is.visualization}
\alias{is.visualization}
\title{Checks if object is a visualization}
\usage{
is.visualization(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Checks if object is a visualization
}
\keyword{internal}
