% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axeslabel.R
\name{is.xlabel}
\alias{is.xlabel}
\title{Checks to see if object is xlabel.}
\usage{
is.xlabel(x)
}
\arguments{
\item{x}{Checks this object.}
}
\description{
Checks to see if object is xlabel.
}
\keyword{internal}
