% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axeslabel.R
\name{is.ylabel}
\alias{is.ylabel}
\title{Checks to see if object is ylabel.}
\usage{
is.ylabel(x)
}
\arguments{
\item{x}{Checks this object.}
}
\description{
Checks to see if object is ylabel.
}
\keyword{internal}
