% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsrange.R
\name{make.lims}
\alias{make.lims}
\title{returns a vector of length 2 containing the limits of the range,
formatted how range expects lims to be formatted}
\usage{
make.lims(x)
}
\arguments{
\item{x}{a vector of any length to be formatted}
}
\description{
returns a vector of length 2 containing the limits of the range,
formatted how range expects lims to be formatted
}
\keyword{internal}
