% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmodel.R
\name{NaNRemove}
\alias{NaNRemove}
\title{NaNRemove}
\usage{
NaNRemove(twoDList)
}
\arguments{
\item{twoDList}{Data of form \code{list(list(),list())}.}
}
\description{
Finds and removes all 2-d points which produce NaNs
}
\keyword{internal}
