% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspoint.R
\name{pnt}
\alias{pnt}
\title{Alias to make constructing points for regions easier.}
\usage{
pnt(x, y, label = "", pch = 21, size = 2, col = "blue",
  regionCol = NULL, image = "", offset = NULL, display = TRUE,
  fixed = FALSE, iters = 0, attractor = FALSE, crop = TRUE,
  artificial = FALSE, ...)
}
\description{
Alias to make constructing points for regions easier.
}
\keyword{internal}
