% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspoint.R
\name{pointsToList}
\alias{pointsToList}
\title{Converts a list of points to a list of x coordinates, y coordinates, region colors, and indexes.}
\usage{
pointsToList(points)
}
\arguments{
\item{points}{A list of dspoints, or anything that supports $x, $y, and $col}
}
\description{
Converts a list of points to a list of x coordinates, y coordinates, region colors, and indexes.
}
\keyword{internal}
