% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmodel.R
\name{safe.apply}
\alias{safe.apply}
\title{Abstract a Function which does not Crash Upon Failure}
\usage{
safe.apply(fun, inp)
}
\arguments{
\item{fun}{Input function to abstract.}

\item{inp}{Input to that function.}
}
\description{
This function takes a boolean function and abstracts it. If the given function fails, the program will not crash.
}
\keyword{internal}
