% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmodel.R
\name{sqdist}
\alias{sqdist}
\title{Returns the squared distance from a to b}
\usage{
sqdist(a, b)
}
\arguments{
\item{a}{a list of points to test}

\item{b}{a list of points to test}
}
\description{
This function takes two list of points returns a list conaining
the squared distance between each pair of points.
}
\keyword{internal}
