% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsproto.R
\docType{data}
\name{visualization}
\alias{visualization}
\title{Visualization}
\format{An object of class \code{visualization} (inherits from \code{facade}, \code{dsproto}) of length 3.}
\usage{
visualization
}
\description{
Visualization
}
\keyword{internal}
