% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axeslabel.R
\name{xlabel}
\alias{xlabel}
\title{xlabel}
\usage{
xlabel(label = "", ...)
}
\arguments{
\item{label}{The title of the axis, to be displayed on image.}

\item{...}{Extra parameters. These parameters are fed into \code{mtext()}.}
}
\description{
Labels the x-axis
}
\examples{
 library(dsmodels)
 fun <- function(X,Y) {
 list(
     X/exp(Y),
     Y/exp(X)
  )
}

model <- dsmodel(fun, title = "Cool Function!")
model + dsrange(-2:3,-2:3, discretize = .09) +
 xlabel("X-Axis!") +
 ylabel("Y-Axis!")
}

